requirements.txt
ollama
groq
python-dotenv
sounddevice
wavio
numpy
keyboard
pyperclip
requests
Pillow
screeninfo
fastapi
uvicorn
pydantic
python-multipart
openai
pystray